==> 4_modeling.ipynb <==
# 4. Modeling

We use the Surprise package to train a matrix-factorization based collaborative filtering recommender system. 

import numpy as np
import pandas as pd
import seaborn as sns
import matplotlib.pyplot as plt
import ast
import pickle
import warnings

import requests
from bs4 import BeautifulSoup
import functools
import operator
import os
from MPAreaTree import MPAreaTree


from surprise import SVD, KNNBasic,KNNWithMeans, SVDpp, NMF, BaselineOnly
from surprise import Dataset, Reader
from surprise import accuracy
from surprise.model_selection import train_test_split, GridSearchCV


pd.set_option('display.max_colwidth', 4000)
warnings.filterwarnings('ignore')

routes = pd.read_csv('./data/routes.csv')
areas = pd.read_csv('./data/areas.csv')
ratings = pd.read_csv('./data/ratings.csv')
area_tree = pickle.load(open('./pickle/area_tree.pkl', 'rb'))


df = ratings[['user_id', 'route_id', 'user_rating']]
reader = Reader(rating_scale = (0,4))
data = Dataset.load_from_df(df, reader = reader)


if False:
    trainset, testset = train_test_split(data, test_size=.25)

    baseline = BaselineOnly()
    svd = SVD()
    svdpp = SVDpp()
    knn = KNNBasic()
    knnwithmeans = KNNWithMeans()
    nmf = NMF()

    for  name, algo in [('Baseline', baseline), ('SVD', svd), ('KNNBasic',knn), 
                 ('KNNWithMeans',knnwithmeans), ('NMF',nmf)]:

        algo.fit(trainset)
        predictions = algo.test(testset)


        print(f'{name}:')
        accuracy.rmse(predictions)

trainset, testset = train_test_split(data, test_size=.25)

ratings[ratings.groupby('user_id')['user_id'].transform('size') >= 50]['user_id'].value_counts()
